requirements.txt
django==5.1.2
django-widget-tweaks==1.5.0
python-dotenv==1.0.1
pillow==11.0.
django-cors-headers==4.6.0
qrcode==8.0
pillow==11.0.0
djangorestframework==3.15.2
pycryptodome==3.21.0
gunicorn==23.0.0
whitenoise==6.8.2
django-storages==1.14.4
boto3==1.35.55